debug/requirements.txt
beautifulsoup4>=4.12.3,<5
lxml>=5.3.0,<6
msgspec>=0.19.0,<1
yarl>=1.18.3,<2
python-dateutil>=2.9.0.post0,<3
rich>=13.9.4,<14

aiohttp[speedups]>=3.11.11,<4
uvloop>=0.21.0,<1 ; sys_platform != "win32"
winloop>=0.1.6,<1 ; sys_platform == "win32"

typer-slim>=0.15.0,<1

ruff>=0.3.4,<1